% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_psych.pipeline.R
\name{configure_psych.pipeline}
\alias{configure_psych.pipeline}
\title{Configure Psychology Pipeline}
\usage{
configure_psych.pipeline(yaml_path)
}
\arguments{
\item{yaml_path}{A string specifying the path to the YAML configuration file.}
}
\value{
A list. The modified YAML data after setting the default values and updating the pipeline.
}
\description{
This function reads a YAML configuration file and sets default parameters for running a psychology pipeline.
The YAML file should contain sections for \code{path}, \code{settings}, and \code{pipeline}. Each section has its own default values
which will be used if not specified in the YAML file. Also, it extracts unique subject identifiers from a directory
or a .csv file depending on the configuration.
}
\details{
The function first checks if the YAML data has \code{path}, \code{settings}, and \code{pipeline} sections. If any section is not
present, it creates the section with default values. If a section is present but does not have all the fields, it
sets the missing fields to their default values.

For the \code{path} section, the default values are \code{pipeline_fx = "R"}, \code{log = "outputs/pipeline_logfiles"}, and
\code{output = "data_preproc"}.

For the \code{settings} section, the default values are \code{project = "untitled project"},
\code{pipeline_description = "psych.pipeline"}, \code{timestamp = TRUE}, \code{n_cores = 1}, and \code{force = FALSE}.

For each element in the \code{pipeline} section, the function sets default values for missing fields. The default values
are \code{depends_on = "none"}, \code{print_output = FALSE}, \code{describe_text = element$func}, \code{arguments = list()},
\code{subject_level = FALSE}, and \code{force = default_settings$force}.
}
\note{
The function extracts unique subject identifiers by looking at filenames in a directory or by reading a .csv file.
It is important to ensure that the filenames or the .csv file contain the subject identifiers.
}
\examples{
\dontrun{
# It is used to set up the pipeline configuration
configure_psych.pipeline(yaml_path = "./inst/dimt_params.yaml")
}

}
\seealso{
\code{\link[yaml]{yaml.load_file}} for loading YAML files.
\code{\link[base]{file.info}}, \code{\link[base]{list.files}} for file and directory operations.
\code{\link[readr]{read_csv}} for reading .csv files.
}
\author{
Nate Hall
}
