% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Internal function to set package options}
\usage{
.onLoad(libname, pkgname)
}
\description{
This function is run when the package is loaded to set various package options.
}
\keyword{internal}
