% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_log.R
\name{execute_log}
\alias{execute_log}
\title{Execute R Code with Error/Warning Logging}
\usage{
execute_log(code, log_path = NULL, output_path = NULL, describe_text = NULL)
}
\arguments{
\item{code}{An R expression or block of code to be executed. Typically a function of an analytic pipeline.}

\item{log_path}{Character string. Path to the log file where errors and warnings are to be written.
If NULL (default), output is written to the console.}

\item{output_path}{Character string. Path to the RDS file where the output of the code is to be stored.
If NULL (default), the output is not stored.}

\item{describe_text}{Character string. A description text to include in the log output.
If NULL (default), no description text is added.}
}
\value{
If there is an error or warning during code execution, a \code{try-error} or \code{simpleWarning} object is returned,
respectively. If the code executes successfully, the output of the code is returned.
}
\description{
This function executes an R expression or block of code, logging errors and warnings.
It optionally logs output to a specified file and stores the output in a RDS file.
}
\examples{
\dontrun{
  execute_log({print("Hello, world!")}, log_path = "my_log.txt")
}

}
\author{
Nate Hall
}
