% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_args.R
\name{expand_args}
\alias{expand_args}
\title{Expand Arguments Based on Run Subjects}
\usage{
expand_args(args, run_subjects)
}
\arguments{
\item{args}{A named list of arguments. The list must include a \code{subj_dat} element
which is the path to the directory containing the data files.}

\item{run_subjects}{A character vector of subject IDs to include.
These IDs should match the base names of the data files (excluding the file extension).}
}
\value{
A list of argument lists, one for each subject in \code{run_subjects}.
Each argument list includes the full path to the subject's data file in the \code{subj_dat} argument,
and all other arguments are copied from \code{args}.
}
\description{
This function takes a list of arguments and a list of subjects to run,
and expands the arguments into a list of argument lists, one for each subject.
The \code{subj_dat} argument is replaced with the full path to the subject's data file.
}
\examples{
args <- list(subj_dat = "data_raw/behav/dimt", arg2 = 2, arg3 = 3)
run_subjects <- c("137", "169", "246", "248", "255", "272", "317", "363", "364", "365", "386", "423", "434", "452", "482", "92")
args_list <- expand_args(args, run_subjects)

}
\author{
Nate Hall
}
