% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subj_ids.R
\name{get_subj_ids}
\alias{get_subj_ids}
\title{Get Subject IDs}
\usage{
get_subj_ids(subject_list)
}
\arguments{
\item{subject_list}{A character string specifying the path to a directory or a .csv file.}
}
\value{
If subject_list is a directory, returns a character vector of file basenames present in the directory without their extensions.
If subject_list is a .csv file, returns a character vector of the contents of the first column in the .csv file.
}
\description{
This function extracts the subject IDs from a specified directory or a .csv file.
If a directory is provided, the function will return a list of file basenames (without extensions) present in the directory.
If a .csv file is provided, the function will read the file and return the first column as a vector.
}
\note{
If the subject_list input is neither a .csv file nor a directory, the function will stop and throw an error.
}
\examples{
\dontrun{
get_subj_ids("/path/to/directory")
get_subj_ids("/path/to/file.csv")
}

}
\seealso{
\code{\link{dir.exists}}, \code{\link{file.exists}}, \code{\link{read.csv}}
}
\author{
Nate Hall
}
\keyword{internal}
