% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_outdirs.R
\name{init_outdirs}
\alias{init_outdirs}
\title{Initialize Output Directories}
\usage{
init_outdirs(config, subj_control)
}
\arguments{
\item{config}{A list containing configuration information. The \verb{$path$output} element should contain the path to the output directory.}

\item{subj_control}{A data frame or tibble containing information about the control of subjects. The \code{subject_level} column should indicate whether the output for each step is subject-level, and the \code{data_destination} and \code{log_destination} columns should contain the output directory paths for each step.}
}
\value{
This function does not return a value, but it will create directories on the filesystem as a side effect.
}
\description{
This function checks for the existence of output directories, and creates them if they do not exist. It is designed to handle both subject-level and non-subject level output directories.
}
\examples{
config <- list(path = list(output = "/path/to/output"))
subj_control <- data.frame(
  subject_level = c(FALSE, TRUE),
  data_destination = c("/path/to/data1", "/path/to/data2"),
  log_destination = c("/path/to/log1", "/path/to/log2")
)
init_outdirs(config, subj_control)

}
\author{
Nate Hall
}
