% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_psych.pipeline.R
\name{launch_psych.pipeline}
\alias{launch_psych.pipeline}
\title{Launch a psych.pipeline instance from a single yaml file.}
\usage{
launch_psych.pipeline(yaml_path = NULL)
}
\arguments{
\item{yaml_path}{The path to the YAML file that contains the configuration settings for the pipeline. If NULL, the most recent YAML file in the current directory is used. Default is NULL.}
}
\value{
The function does not return a value; it launches a single psych.pipeline call based on the parameters.
}
\description{
This function launches the psych.pipeline either on a local machine or a high-performance computing cluster.
}
\author{
Nate Hall
}
