% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tree.R
\name{list_tree}
\alias{list_tree}
\title{Print a tree-like structure of a list}
\usage{
list_tree(x, prefix = "", is_tail = TRUE)
}
\arguments{
\item{x}{A list to print. This list can be nested (i.e., contain other lists).}

\item{prefix}{A string used to control the indentation of the tree structure.}

\item{is_tail}{A logical value indicating whether the current item is the last in its parent list.}
}
\value{
Invisibly returns the original list, \code{x}, and primarily operates through side effects (printing to the console).
}
\description{
This function prints a tree-like structure of a list to the console. It handles nested lists,
and uses special characters to represent the branching of the tree.
}
\examples{
\dontrun{
x <- list(
  DESCRIPTION = "Package description",
  NAMESPACE = "Package namespace",
  R = list(
    function1 = "Code for function 1",
    function2 = "Code for function 2"
  )
)
list_tree(x)

}


}
