% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_wide.R
\name{print_wide}
\alias{print_wide}
\title{Print a tibble without restricting width}
\usage{
print_wide(tib)
}
\arguments{
\item{tib}{A tibble to be printed.}
}
\description{
This function prints a tibble with all columns visible.
}
\examples{
# Assuming df is a tibble
print_wide(df)

}
