% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_directory.R
\name{source_directory}
\alias{source_directory}
\title{Source all R scripts in a directory}
\usage{
source_directory(dir_path, recursive = TRUE)
}
\arguments{
\item{dir_path}{A character string specifying the path to the directory
containing the R scripts to be sourced.}

\item{recursive}{list recurisvely into the directory being sourced}
}
\value{
This function has no return value.
}
\description{
This function sources all the R scripts in a given directory by calling the
source() function on each file with a ".R" extension. This function can be
used to easily source all the scripts in a directory.
}
\examples{
\dontrun{
# Source all R scripts in the current working directory
source_directory(getwd())
}

}
\author{
Nate Hall
}
